deploy_deit_aws_ec2/fast_api_app/requirements.txt
fastapi==0.78.0
gunicorn==20.1.0
numpy==1.22.1
Pillow==9.1.1
pydantic==1.9.1
requests==2.27.1
transformers==4.19.4
uvicorn==0.17.6
